# Zserver.web/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
    "title": "go-admin API",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/wenjianzhang/ZserverWeb/blob/master/LICENSE.md"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/db/tables/page": {
      "get": {
        "description": "数据库表分页列表 / database table page list",
        "tags": [
          "工具 / Tools"
        ],
        "summary": "分页列表数据 / page list data",
        "parameters": [
          {
            "type": "string",
            "description": "tableName / 数据表名称",
            "name": "tableName",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "pageSize / 页条数",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "pageIndex / 页码",
            "name": "pageIndex",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/dict/type": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "字典类型"
        ],
        "summary": "修改字典类型",
        "parameters": [
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.DictType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "字典类型"
        ],
        "summary": "添加字典类型",
        "parameters": [
          {
            "description": "data",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.DictType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/dict/type/list": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "字典类型"
        ],
        "summary": "字典类型列表数据",
        "parameters": [
          {
            "type": "string",
            "description": "dictName",
            "name": "dictName",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictId",
            "name": "dictId",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictType",
            "name": "dictType",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页条数",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "pageIndex",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
    "/api/v1/dict/type/{dictId}": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "字典类型"
        ],
        "summary": "通过字典id获取字典类型",
        "parameters": [
          {
            "type": "integer",
            "description": "字典类型编码",
            "name": "dictId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      },
      "delete": {
        "description": "删除数据",
        "tags": [
          "字典类型"
        ],
        "summary": "删除字典类型",
        "parameters": [
          {
            "type": "integer",
            "description": "dictId",
            "name": "dictId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"删除失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "app.Response": {
      "type": "object",
      "properties": {
        "code": {
          "description": "代码",
          "type": "integer",
          "example": 200
        },
        "data": {
          "description": "数据集",
          "type": "object"
        },
        "msg": {
          "description": "消息",
          "type": "string"
        }
      }
    },
...
    "models.DictType": {
      "type": "object",
      "properties": {
        "createBy": {
          "description": "创建者",
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "dataScope": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "dictId": {
          "type": "integer"
        },
        "dictName": {
          "description": "字典名称",
          "type": "string"
        },
        "dictType": {
          "description": "字典类型",
          "type": "string"
        },
        "params": {
          "type": "string"
        },
        "remark": {
          "description": "备注",
          "type": "string"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "updateBy": {
          "description": "更新者",
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}